Move rendering of HomeDashBoardMyProfile to MainContent instead of in App
Positioning improved, but still not satisfactory
